# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Gemini Rent Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /properties:
    get:
      operationId: listProperties
      tags: [properties]
      summary: List all properties
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of properties
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Property"
    post:
      operationId: createProperty
      tags: [properties]
      summary: Create a new property
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PropertyInput"
      responses:
        "201":
          description: Property created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
  /properties/{id}:
    get:
      operationId: getProperty
      tags: [properties]
      summary: Get a property by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Property details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
        "404":
          description: Not found
    patch:
      operationId: updateProperty
      tags: [properties]
      summary: Update a property
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PropertyUpdate"
      responses:
        "200":
          description: Property updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
    delete:
      operationId: deleteProperty
      tags: [properties]
      summary: Delete a property
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /tenants:
    get:
      operationId: listTenants
      tags: [tenants]
      summary: List all tenants
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: propertyId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of tenants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tenant"
    post:
      operationId: createTenant
      tags: [tenants]
      summary: Create a new tenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantInput"
      responses:
        "201":
          description: Tenant created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
  /tenants/{id}:
    get:
      operationId: getTenant
      tags: [tenants]
      summary: Get a tenant by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tenant details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "404":
          description: Not found
    patch:
      operationId: updateTenant
      tags: [tenants]
      summary: Update a tenant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantUpdate"
      responses:
        "200":
          description: Tenant updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
    delete:
      operationId: deleteTenant
      tags: [tenants]
      summary: Delete a tenant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List all rent payments
      parameters:
        - name: tenantId
          in: query
          schema:
            type: integer
        - name: propertyId
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Record a rent payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Payment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}:
    get:
      operationId: getPayment
      tags: [payments]
      summary: Get a payment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "404":
          description: Not found
...
  /reports/monthly:
    get:
      operationId: getMonthlyReport
      tags: [reports]
      summary: Get monthly income/expense report
      parameters:
        - name: year
          in: query
          required: true
          schema:
            type: integer
        - name: month
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Monthly report data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyReport"
  /reports/yearly:
    get:
      operationId: getYearlyReport
      tags: [reports]
      summary: Get yearly income/expense report
      parameters:
        - name: year
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Yearly report data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/YearlyReport"
components:
  schemas:
...
    Property:
      type: object
      required: [id, name, address, type, totalUnits, rentAmount, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: string
        type:
          type: string
          enum: [apartment, house, commercial, land]
        totalUnits:
          type: integer
        rentAmount:
          type: number
        status:
          type: string
          enum: [available, occupied, maintenance]
        description:
          type: ["string", "null"]
        createdAt:
          type: string
    PropertyInput:
      type: object
      required: [name, address, type, totalUnits, rentAmount]
      properties:
        name:
          type: string
        address:
          type: string
        type:
          type: string
          enum: [apartment, house, commercial, land]
        totalUnits:
          type: integer
        rentAmount:
          type: number
        status:
          type: string
          enum: [available, occupied, maintenance]
        description:
          type: string
    PropertyUpdate:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        type:
          type: string
          enum: [apartment, house, commercial, land]
        totalUnits:
          type: integer
        rentAmount:
          type: number
        status:
          type: string
          enum: [available, occupied, maintenance]
        description:
          type: string
    Tenant:
      type: object
      required: [id, name, email, phone, propertyId, unitNumber, leaseStart, leaseEnd, rentAmount, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        propertyId:
          type: integer
        propertyName:
          type: ["string", "null"]
        unitNumber:
          type: string
        leaseStart:
          type: string
        leaseEnd:
          type: string
        rentAmount:
          type: number
        status:
          type: string
          enum: [active, inactive, evicted]
        emergencyContact:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    TenantInput:
      type: object
      required: [name, email, phone, propertyId, unitNumber, leaseStart, leaseEnd, rentAmount]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        propertyId:
          type: integer
        unitNumber:
          type: string
        leaseStart:
          type: string
        leaseEnd:
          type: string
        rentAmount:
          type: number
        status:
          type: string
          enum: [active, inactive, evicted]
        emergencyContact:
          type: string
        notes:
          type: string
    TenantUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        propertyId:
          type: integer
        unitNumber:
          type: string
        leaseStart:
          type: string
        leaseEnd:
          type: string
        rentAmount:
          type: number
        status:
          type: string
          enum: [active, inactive, evicted]
        emergencyContact:
          type: string
        notes:
          type: string
    Payment:
      type: object
      required: [id, tenantId, propertyId, amount, paymentDate, month, year, method, status, createdAt]
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        tenantName:
          type: ["string", "null"]
        propertyId:
          type: integer
        propertyName:
          type: ["string", "null"]
        amount:
          type: number
        paymentDate:
          type: string
        month:
          type: integer
        year:
          type: integer
        method:
          type: string
          enum: [cash, bank_transfer, upi, cheque, online]
        status:
          type: string
          enum: [paid, pending, partial, overdue]
        notes:
          type: ["string", "null"]
        receiptNumber:
          type: ["string", "null"]
        createdAt:
          type: string
    PaymentInput:
      type: object
      required: [tenantId, propertyId, amount, paymentDate, month, year, method]
      properties:
        tenantId:
          type: integer
        propertyId:
          type: integer
        amount:
          type: number
        paymentDate:
          type: string
        month:
          type: integer
        year:
          type: integer
        method:
          type: string
          enum: [cash, bank_transfer, upi, cheque, online]
        status:
          type: string
          enum: [paid, pending, partial, overdue]
        notes:
          type: string
    Expense:
      type: object
      required: [id, title, amount, category, date, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        amount:
          type: number
        category:
          type: string
          enum: [repair, utility, tax, insurance, maintenance, salary, other]
        date:
          type: string
        propertyId:
          type: ["integer", "null"]
        propertyName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
...
    MonthlyReport:
      type: object
      required: [year, month, totalIncome, totalExpenses, netProfit, payments, expenses]
      properties:
        year:
          type: integer
        month:
          type: integer
        totalIncome:
          type: number
        totalExpenses:
          type: number
        netProfit:
          type: number
        collectedRents:
          type: integer
        pendingRents:
          type: integer
        payments:
          type: array
          items:
            $ref: "#/components/schemas/Payment"
        expenses:
          type: array
          items:
            $ref: "#/components/schemas/Expense"
    YearlyReport:
      type: object
      required: [year, totalIncome, totalExpenses, netProfit, monthlyBreakdown]
      properties:
        year:
          type: integer
        totalIncome:
          type: number
        totalExpenses:
          type: number
        netProfit:
          type: number
        monthlyBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/MonthlyBreakdown"
    MonthlyBreakdown:
      type: object
      required: [month, income, expenses, netProfit]
      properties:
        month:
          type: integer
        income:
          type: number
        expenses:
          type: number
        netProfit:
          type: number