# lib/api-zod/src/generated/types/property.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Gemini Rent Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { PropertyStatus } from './propertyStatus';
import type { PropertyType } from './propertyType';

export interface Property {
  id: number;
  name: string;
  address: string;
  type: PropertyType;
  totalUnits: number;
  rentAmount: number;
  status: PropertyStatus;
  /** @nullable */
  description?: string | null;
  createdAt: string;
}