# lib/api-zod/src/generated/types/propertyUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Gemini Rent Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { PropertyUpdateStatus } from './propertyUpdateStatus';
import type { PropertyUpdateType } from './propertyUpdateType';

export interface PropertyUpdate {
  name?: string;
  address?: string;
  type?: PropertyUpdateType;
  totalUnits?: number;
  rentAmount?: number;
  status?: PropertyUpdateStatus;
  description?: string;
}